+++ vim/dotfiles/.config/nvim/lua/pelyib/gitblame.lua
@@ -1,0 +1,7 @@
+local M = {}
+
+M.setup = function ()
+    vim.keymap.set('n', '<leader>gb', '<cmd>GitBlameToggle<cr>', {})
+end
+
+return M