+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.12] - 2023.05.29
+
+### Added
+- [NVIM] - before and after local modules around the `pelyib` modules
+
+### Changed
+- [NVIM] - introduce `pelyib` lua module, move and split the init.lua to specific modules
+
+### Removed
+- [NVIM] - init.lua.dist, use before and after local modules instead 
+
 ## [0.0.11] - 2023.05.23
 
 ### Changed